requirements.txt
Flask==3.0.3
boto3
gunicorn==20.1.0
Werkzeug==3.0.6
pdf2image==1.17.0
flask-swagger-ui==4.11.1
python-dotenv==1.0.1
Pillow==10.3.0
openai==1.44.1
langchain-openai==0.1.23
langchain==0.2.16
langchain-community==0.2.4
google-cloud-documentai==2.9.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-core==2.11.0
pytest
pytest-cov
pytest-mock